requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_openmp_mutex=4.5=2_gnu
absl-py=2.1.0=py312haa95532_0
aiohappyeyeballs=2.4.0=py312haa95532_0
aiohttp=3.10.5=py312h827c3e9_0
aiosignal=1.2.0=pyhd3eb1b0_0
asttokens=3.0.0=pyhd8ed1ab_1
attrs=24.2.0=py312haa95532_0
aws-c-auth=0.8.0=h75ad88d_4
aws-c-cal=0.8.0=h0da4a7a_0
aws-c-common=0.9.31=h2466b09_0
aws-c-compression=0.3.0=h0da4a7a_0
aws-c-event-stream=0.5.0=h3ba0563_2
aws-c-http=0.9.0=h4d69eff_3
aws-c-io=0.15.0=hbd1f77e_2
aws-c-mqtt=0.11.0=h5d66fae_2
aws-c-s3=0.7.0=h7a7f505_5
aws-c-sdkutils=0.2.0=h0da4a7a_0
aws-checksums=0.2.0=h0da4a7a_0
aws-crt-cpp=0.29.0=h04d40f5_6
aws-sdk-cpp=1.11.407=h35367fc_6
blas=1.0=mkl
brotli=1.0.9=ha925a31_2
brotli-python=1.0.9=py312hd77b12b_8
bzip2=1.0.8=h2bbff1b_6
c-ares=1.34.2=h2466b09_0
ca-certificates=2024.12.31=haa95532_0
certifi=2024.12.14=py312haa95532_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
colorama=0.4.6=py312haa95532_0
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.0=py312hd5eb7cc_2
cuda-cccl=12.6.77=0
cuda-cccl_win-64=12.6.77=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.6.77=0
cuda-runtime=11.8.0=0
cuda-version=12.6=3
cycler=0.11.0=pyhd3eb1b0_0
cython=3.0.11=py312h827c3e9_0
datasets=3.0.2=pyhd8ed1ab_0
debugpy=1.8.11=py312h275cf98_0
decorator=5.1.1=pyhd8ed1ab_1
dill=0.3.8=py312haa95532_0
einops=0.8.0=pyhd8ed1ab_1
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_1
expat=2.6.3=h5da7b33_0
filelock=3.13.1=py312haa95532_0
fonttools=4.51.0=py312h2bbff1b_0
freetype=2.12.1=hdaf720e_2
frozenlist=1.4.0=py312h2bbff1b_0
fsspec=2024.6.1=py312haa95532_0
giflib=5.2.2=h7edc060_0
grpcio=1.67.1=py312h1f7db74_0
huggingface_hub=0.24.6=py312haa95532_0
icc_rt=2022.1.0=h6049295_2
idna=3.7=py312haa95532_0
importlib-metadata=8.5.0=pyha770c72_1
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.29.5=pyh4bbf305_0
ipython=8.31.0=pyh7428d3b_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.4=py312haa95532_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=py312h2e8e312_0
kiwisolver=1.4.4=py312hd77b12b_0
krb5=1.21.3=hdf4eb48_0
lcms2=2.16=h67d730c_0
lerc=4.0.0=h63175ca_0
libabseil=20240722.0=cxx17_he0c23c2_1
libarrow=18.0.0=hc3f8aa2_1_cuda
libarrow-acero=18.0.0=hac47afa_1_cuda
libarrow-dataset=18.0.0=hac47afa_1_cuda
libarrow-substrait=18.0.0=hcd1cebd_1_cuda
libbrotlicommon=1.1.0=h2466b09_2
libbrotlidec=1.1.0=h2466b09_2
libbrotlienc=1.1.0=h2466b09_2
libcrc32c=1.1.2=hd77b12b_0
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.7.77=0
libcurand-dev=10.3.7.77=0
libcurl=8.10.1=h1ee3ff0_0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.22=h2466b09_0
libevent=2.1.12=h56d1f94_1
libexpat=2.6.3=he0c23c2_0
libffi=3.4.4=hd77b12b_1
libgcc=14.2.0=h1383e82_1
libgomp=14.2.0=h1383e82_1
libgoogle-cloud=2.30.0=h07d40e7_1
libgoogle-cloud-storage=2.30.0=he5eb982_1
libgrpc=1.67.1=h7aa3b8a_0
libjpeg-turbo=3.0.3=h827c3e9_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libparquet=18.0.0=h59f2d37_1_cuda
libpng=1.6.44=h3ca93ac_0
libprotobuf=5.28.2=hcaed137_0
libre2-11=2024.07.02=h4eb7d71_1
libsodium=1.0.20=hc70643c_0
libsqlite=3.47.0=h2466b09_1
libssh2=1.11.0=h7dfc565_0
libthrift=0.21.0=hbe90ef8_0
libtiff=4.7.0=hfc51747_1
libutf8proc=2.8.0=h82a8f57_0
libuv=1.48.0=h827c3e9_0
libwebp=1.4.0=h2466b09_0
libwebp-base=1.4.0=hcfcfb64_0
libwinpthread=12.0.0.r4.gg4f2fc60ca=h57928b3_8
libxcb=1.17.0=h0e4246c_0
libzlib=1.3.1=h2466b09_2
lz4-c=1.9.4=h2bbff1b_1
markdown=3.4.1=py312haa95532_0
markupsafe=2.1.3=py312h2bbff1b_0
matplotlib-base=3.9.2=py312h90004f6_1
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py312h2bbff1b_1
mkl_fft=1.3.10=py312h827c3e9_0
mkl_random=1.2.7=py312h0158946_0
mpmath=1.3.0=py312haa95532_0
multidict=6.0.4=py312h2bbff1b_0
multiprocess=0.70.15=py312haa95532_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.2.1=py312haa95532_0
numpy=1.26.4=py312hfd52020_0
numpy-base=1.26.4=py312h4dde369_0
openjpeg=2.5.2=hae555c5_0
openssl=3.4.0=ha4e3fda_1
orc=2.0.2=h34659fe_2
packaging=24.1=py312haa95532_0
pandas=2.2.3=py312h72972c8_1
parso=0.8.4=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.0.0=py312ha41cd45_0
pip=24.2=py312haa95532_0
platformdirs=4.3.6=pyhd8ed1ab_1
prompt-toolkit=3.0.48=pyha770c72_1
protobuf=5.28.2=py312h275cf98_0
psutil=6.1.1=py312h4389bb4_0
pthread-stubs=0.3=h3c9f919_1
pure_eval=0.2.3=pyhd8ed1ab_1
pyarrow=18.0.0=py312h7e22eef_0
pyarrow-core=18.0.0=py312h607bf26_0_cuda
pybind11-abi=5=hd3eb1b0_0
pycocotools=2.0.8=py312h1a27103_0
pygments=2.19.1=pyhd8ed1ab_0
pyparsing=3.1.2=py312haa95532_0
pysocks=1.7.1=py312haa95532_0
python=3.12.7=hce54a09_0_cpython
python-dateutil=2.9.0post0=py312haa95532_2
python-tzdata=2023.3=pyhd3eb1b0_0
python-xxhash=2.0.2=py312h2bbff1b_1
python_abi=3.12=5_cp312
pytorch=2.5.1=py3.12_cuda11.8_cudnn9_0
pytorch-cuda=11.8=h24eeafa_6
pytorch-mutex=1.0=cuda
pytz=2024.1=py312haa95532_0
pywin32=307=py312h275cf98_3
pyyaml=6.0.2=py312h827c3e9_0
pyzmq=26.2.0=py312hd7027bb_3
qhull=2020.2=h59b6b97_2
re2=2024.07.02=hd3b24a8_1
regex=2024.9.11=py312h827c3e9_0
requests=2.32.3=py312haa95532_0
safetensors=0.4.5=py312h1429478_0
scipy=1.13.1=py312hbb039d4_0
seaborn=0.13.2=py312haa95532_0
setuptools=72.1.0=py312haa95532_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.2.1=hcdb6601_0
sqlite=3.45.3=h2bbff1b_0
stack_data=0.6.3=pyhd8ed1ab_1
sympy=1.13.2=py312haa95532_0
tbb=2021.8.0=h59b6b97_0
tensorboard=2.18.0=pyhd8ed1ab_0
tensorboard-data-server=0.7.0=py312haa95532_1
timm=1.0.11=pyhd8ed1ab_0
tk=8.6.13=h5226925_1
tokenizers=0.20.1=py312h4b3af92_0
torchaudio=2.5.1=pypi_0
torchvision=0.20.1=pypi_0
tornado=6.4.2=py312h4389bb4_0
tqdm=4.66.5=py312hfc267ef_0
traitlets=5.14.3=pyhd8ed1ab_1
transformers=4.46.1=pyhd8ed1ab_0
typing-extensions=4.11.0=py312haa95532_0
typing_extensions=4.11.0=py312haa95532_0
tzdata=2024b=h04d1e81_0
ucrt=10.0.20348.0=haa95532_0
unicodedata2=15.1.0=py312h2bbff1b_0
urllib3=2.2.3=py312haa95532_0
vc=14.40=h2eaa2aa_1
vc14_runtime=14.40.33810=hcc2c482_22
vs2015_runtime=14.40.33810=h3bf8584_22
wcwidth=0.2.13=pyhd8ed1ab_1
werkzeug=3.0.6=py312haa95532_0
wheel=0.44.0=py312haa95532_0
win_inet_pton=1.1.0=py312haa95532_0
xorg-libxau=1.0.11=h0e40799_1
xorg-libxdmcp=1.1.5=h0e40799_0
xxhash=0.8.0=h2bbff1b_3
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
yarl=1.11.0=py312h827c3e9_0
zeromq=4.3.5=ha9f60a1_7
zipp=3.21.0=pyhd8ed1ab_1
zstd=1.5.6=h0ea2cb4_0
